# docs/openapi.yaml
openapi: 3.1.0
info:
  title: FoodHelper API
  version: v1
  description: API for managing products, recipes, and nutritional calculations.
servers:
  - url: /
    description: Current server
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/planning:
    get:
      tags:
        - Planning
      summary: List planning
      description: Returns compact planning summaries ordered by start date descending, without days, products, or calculated totals.
      operationId: listPlanning
      responses:
        '200':
          description: Planning summaries returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PlanningSummaryResponse'
    post:
      tags:
        - Planning
      summary: Create planning
      description: Starts an empty planning with an inclusive date range. The range may cover up to 16 calendar days, and planning can contain only the days planned so far.
      operationId: createPlanning
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlanningRequest'
            examples:
              default:
                value:
                  startDate: 2026-06-15
                  endDate: 2026-06-30
      responses:
        '201':
          description: Planning created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanningResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v1/planning/{id}:
    get:
      tags:
        - Planning
      summary: Get planning
      description: Returns one planning with ordered days, sections, products, nutritional totals, and a stock preview summary.
      operationId: getPlanning
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Planning returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanningResponse'
        '404':
          description: Planning not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v1/planning/{id}/days:
    put:
      tags:
        - Planning
      summary: Create or replace planned day
      description: Creates or replaces one planned day. Sections and products keep the explicit order supplied in the request, and product sortOrder values must be unique within each section.
      operationId: upsertPlanningDay
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertPlannedDayRequest'
            examples:
              default:
                value:
                  date: 2026-06-15
                  sections:
                    - dayPartId: 1
                      products:
                        - productId: 1
                          units: 1
                          grams: 200
                          sortOrder: 10
                        - productId: 3
                          units: 1
                          sortOrder: 20
                    - dayPartId: 2
                      products:
                        - productId: 2
                          units: 2
                          sortOrder: 10
      responses:
        '200':
          description: Planning updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanningResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Planning or product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v1/planning/{id}/menu:
    post:
      tags:
        - Planning
      summary: Create menu from planning
      description: Consumes the user-confirmed stock allocation, or automatically uses the earliest expiration date when no allocation is supplied. It creates a menu snapshot, stores missing products as a shopping list, and subtracts the applied stock cost from the selected payer user's money box.
      operationId: createMenuFromPlanning
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMenuRequest'
            examples:
              default:
                value:
                  payerUserId: 1
                  stockAllocations:
                    - stockEntryId: 7
                      usedUnits: 1.5
      responses:
        '201':
          description: Menu created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Planning or payer user not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /api/v1/menus/{id}:
    get:
      tags:
        - Menus
      summary: Get menu
      description: Returns one menu snapshot with ordered days, nutritional totals, rule evaluation, stock summary, consumed stock, and missing products.
      operationId: getMenu
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Menu returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuResponse'
        '404':
          description: Menu not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    delete:
      tags:
        - Menus
      summary: Undo menu creation
      description: Returns every consumed quantity to its original stock entry, removes the linked money movement, and deletes the open menu so its planning can be established again. Closed menus cannot be undone.
      operationId: undoMenuCreation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Menu creation undone
        '400':
          description: Closed menu cannot be undone
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Menu not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  schemas:
    ApiError:
      type: object
      properties:
        status:
          type: integer
          example: 400
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: Invalid request body
        path:
          type: string
          example: /api/v1/products
        timestamp:
          type: string
          format: date-time
...
    NutrientRuleEvaluationResponse:
      type: object
      properties:
        value:
          type: number
          description: Observed average per planned day
        minimum:
          type: number
          nullable: true
        maximum:
          type: number
          nullable: true
        status:
          type: string
          enum: [BELOW_MINIMUM, WITHIN_RANGE, ABOVE_MAXIMUM, NOT_CONFIGURED]
    NutritionalRulesEvaluationResponse:
      type: object
      properties:
        plannedDays:
          type: integer
          format: int32
        calories:
          $ref: '#/components/schemas/NutrientRuleEvaluationResponse'
        carbohydrates:
          $ref: '#/components/schemas/NutrientRuleEvaluationResponse'
        proteins:
          $ref: '#/components/schemas/NutrientRuleEvaluationResponse'
        fats:
          $ref: '#/components/schemas/NutrientRuleEvaluationResponse'
    CreatePlanningRequest:
      description: Payload for starting an empty planning. The inclusive date range cannot span more than 16 calendar days.
      type: object
      required:
        - startDate
        - endDate
      properties:
        startDate:
          type: string
          format: date
          description: First date covered by the planning
          example: 2026-06-15
        endDate:
          type: string
          format: date
          description: Last date covered by the planning
          example: 2026-06-30
    CreateMenuRequest:
      description: Request used to establish a planning, assign its cost, and optionally confirm the exact stock allocation.
      type: object
      required:
        - payerUserId
      properties:
        payerUserId:
          type: integer
          format: int64
          description: User identifier that assumes the menu period cost
          example: 1
        stockAllocations:
          type: array
          nullable: true
          description: Optional user-confirmed stock allocation. When omitted, stock is allocated automatically by earliest expiration date. An empty list consumes no stock.
          items:
            $ref: '#/components/schemas/MenuStockAllocationRequest'
    MenuStockAllocationRequest:
      type: object
      required:
        - stockEntryId
        - usedUnits
      properties:
        stockEntryId:
          type: integer
          format: int64
          description: Stock entry identifier
          example: 7
        usedUnits:
          type: number
          minimum: 0.01
          description: Quantity to consume from this stock entry
          example: 1.5
    UpsertPlannedDayRequest:
      type: object
      required:
        - date
        - sections
      properties:
        date:
          type: string
          format: date
          description: Date of the planned day
          example: 2026-06-15
        sections:
          type: array
          minItems: 1
          description: Configured day parts selected for the day. Each day part can appear only once.
          items:
            $ref: '#/components/schemas/PlanningSectionRequest'
    PlanningSectionRequest:
      type: object
      required:
        - dayPartId
        - products
      properties:
        dayPartId:
          type: integer
          format: int64
          description: Reusable day part identifier
          example: 1
        products:
          type: array
          minItems: 1
          description: Ordered products consumed in this section. Each product sortOrder must be unique within the section.
          items:
            $ref: '#/components/schemas/PlanningProductRequest'
    PlanningProductRequest:
      type: object
      required:
        - productId
        - sortOrder
      properties:
        productId:
          type: integer
          format: int64
          description: Product identifier
          example: 1
        units:
          type: number
          exclusiveMinimum: 0
          nullable: true
          description: Consumed units. Defaults to 1 when omitted.
          example: 2
        grams:
          type: number
          exclusiveMinimum: 0
          nullable: true
          description: Consumed grams. Defaults to product gramsPerUnit multiplied by units when omitted.
          example: 300
        sortOrder:
          type: integer
          format: int32
          minimum: 0
          description: Explicit display order within the section. Must be unique among products in the same section.
          example: 10
    PlanningSummaryResponse:
      type: object
      required: [id, startDate, endDate, plannedDays, state]
      properties:
        id:
          type: integer
          format: int64
          example: 42
        startDate:
          type: string
          format: date
          example: 2026-06-15
        endDate:
          type: string
          format: date
          example: 2026-06-21
        plannedDays:
          type: integer
          format: int32
          minimum: 0
          description: Number of explicitly planned days
          example: 5
        state:
          type: string
          enum: [DRAFT, ESTABLISHED, CLOSED]
          example: ESTABLISHED
        menuId:
          type: [integer, 'null']
          format: int64
          description: Established menu identifier, or null while planning is a draft
          example: 18
    PlanningResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Planning identifier
          example: 1
        startDate:
          type: string
          format: date
          description: First date covered by the planning
          example: 2026-06-15
        endDate:
          type: string
          format: date
          description: Last date covered by the planning
          example: 2026-06-22
        days:
          type: array
          description: Existing planned days. Planning may start empty.
          items:
            $ref: '#/components/schemas/PlanningDayResponse'
        nutritionalValues:
          $ref: '#/components/schemas/NutritionalValuesResponse'
        stockSummary:
          $ref: '#/components/schemas/PlanningStockSummaryResponse'
        nutritionalRules:
          $ref: '#/components/schemas/NutritionalRulesEvaluationResponse'
    MenuResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Current menu identifier
          example: 1
        planningId:
          type: integer
          format: int64
          description: Planning identifier used to create the menu
          example: 5
        payerUserId:
          type: integer
          format: int64
          description: User identifier that assumes the menu period cost
          example: 1
        payerUsername:
          type: string
          description: Username that assumes the menu period cost
          example: elias
        startDate:
          type: string
          format: date
          description: First date covered by the menu period
          example: 2026-06-15
        endDate:
          type: string
          format: date
          description: Last date covered by the menu period
          example: 2026-06-22
        days:
          type: array
          description: Saved day menus in the menu period
          items:
            $ref: '#/components/schemas/PlanningDayResponse'
        nutritionalValues:
          $ref: '#/components/schemas/NutritionalValuesResponse'
        stockSummary:
          $ref: '#/components/schemas/PlanningStockSummaryResponse'
        usedStock:
          type: array
          description: Stock entries consumed to satisfy the menu period
          items:
            $ref: '#/components/schemas/MenuUsedStockResponse'
        shoppingList:
          type: array
          description: Products still missing after consuming stock
          items:
            $ref: '#/components/schemas/MenuShoppingListItemResponse'
        nutritionalRules:
          $ref: '#/components/schemas/NutritionalRulesEvaluationResponse'
...
    PlanningStockSummaryResponse:
      type: object
      properties:
        plannedDays:
          type: integer
          format: int32
          description: Days in the menu that already have at least one saved day
          example: 3
        distinctProducts:
          type: integer
          format: int32
          description: Distinct products used across all planned days
          example: 8
        calories:
          $ref: '#/components/schemas/PlanningStockSummaryCaloriesResponse'
        estimatedCost:
          type: number
          description: Total estimated cost covered by available stock
          example: 18.5
        requirements:
          type: array
          description: Per-product stock coverage requirements
          items:
            $ref: '#/components/schemas/PlanningStockRequirementResponse'
    PlanningStockSummaryCaloriesResponse:
      type: object
      properties:
        averagePerPlannedDay:
          type: number
          description: Average calories per planned day
          example: 1234.56
        maxDay:
          $ref: '#/components/schemas/PlanningStockSummaryDayCaloriesResponse'
        minDay:
          $ref: '#/components/schemas/PlanningStockSummaryDayCaloriesResponse'
    PlanningStockSummaryDayCaloriesResponse:
      type: object
      properties:
        date:
          type: string
          format: date
          description: Day date
          example: 2026-06-15
        calories:
          type: number
          description: Calories for the day
          example: 1450.25
    PlanningStockRequirementResponse:
      type: object
      properties:
        productId:
          type: integer
          format: int64
          description: Product identifier
          example: 1
        productName:
          type: string
          description: Product name
          example: Apple
        requiredUnits:
          type: number
          description: Total units required by the planning
          example: 4.5
        availableUnits:
          type: number
          description: Total units available in stock
          example: 6
        coveredUnits:
          type: number
          description: Units that can be covered by available stock
          example: 4.5
        missingUnits:
          type: number
          description: Units still missing after consuming all available stock
          example: 0
        estimatedCost:
          type: number
          description: Estimated cost for the covered units
          example: 12.75
    PlanningDayResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Planned day identifier
          example: 1
        date:
          type: string
          format: date
          description: Date of the planned day
          example: 2026-06-15
        sections:
          type: array
          description: Configured day parts selected for the day
          items:
            $ref: '#/components/schemas/PlanningSectionResponse'
        nutritionalValues:
          $ref: '#/components/schemas/NutritionalValuesResponse'
    PlanningSectionResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Section identifier
          example: 1
        dayPartId:
          type: integer
          format: int64
          description: Reusable day part identifier
          example: 1
        name:
          type: string
          description: Section name
          example: Lunch
        description:
          type: string
          description: Section description copied from the reusable day part
          example: Main meal of the day
        sortOrder:
          type: integer
          format: int32
          description: Explicit display order within the day
          example: 10
        products:
          type: array
          description: Ordered products consumed in this section
          items:
            $ref: '#/components/schemas/PlanningProductResponse'
        nutritionalValues:
          $ref: '#/components/schemas/NutritionalValuesResponse'
...
    PlanningProductResponse:
      type: object
      properties:
        productId:
          type: integer
          format: int64
          description: Product identifier
          example: 1
        productName:
          type: string
          description: Product name
          example: Apple
        units:
          type: number
          description: Consumed units
          example: 2
        grams:
          type: number
          description: Consumed grams
          example: 300
        sortOrder:
          type: integer
          format: int32
          description: Explicit display order within the section. Values are unique within the section.
          example: 10
        nutritionalValues:
          $ref: '#/components/schemas/NutritionalValuesResponse'
    MenuUsedStockResponse:
      type: object
      properties:
        stockEntryId:
          type: integer
          format: int64
          description: Stock entry identifier
          example: 7
        productId:
          type: integer
          format: int64
          description: Product identifier
          example: 1
        productName:
          type: string
          description: Product name
          example: Apple
        usedUnits:
          type: number
          description: Quantity consumed from the stock entry
          example: 1.5
        price:
          type: number
          description: Price paid for one unit in this stock entry
          example: 2.49
        totalCost:
          type: number
          description: Total cost of the consumed quantity
          example: 3.74
        expirationDate:
          type: string
          format: date
          nullable: true
          description: Stock expiration date when available
          example: 2026-06-20
        entryDate:
          type: string
          format: date
          description: Date when the stock entry was received
          example: 2026-06-10
    MenuShoppingListItemResponse:
      type: object
      properties:
        productId:
          type: integer
          format: int64
          description: Product identifier
          example: 1
        productName:
          type: string
          description: Product name
          example: Apple
        missingUnits:
          type: number
          description: Units still missing after consuming available stock
          example: 2
    NutritionalValuesResponse:
      type: object
      properties:
        calories:
          type: number
          description: Calories
          example: 52
        carbohydrates:
          type: number
          description: Carbohydrates
          example: 14
        proteins:
          type: number
          description: Proteins
          example: 0.3
        fats:
          type: number
          description: Fats
          example: 0.2